+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Correctly compare resolve rule and associations
+
 ## [0.4.4] - 2022-11-25
 
 ### Fixed